fixd pokemon delete API route / controller to match restful conventions

+++ server/routes/api.php
@@ -26,7 +26,7 @@ Route::get('/pokemons/region/{region}', [PokemonController::class, 'list']);
 Route::get('/pokemons/{id}', [PokemonController::class, 'find'])->where('id', '[0-9]+');
 
 Route::post('/pokemons/add', [PokemonController::class, 'addToFavorite']);
-Route::delete('/pokemons/remove', [PokemonController::class, 'deleteFromFavorite']);
+Route::delete('/pokemons/remove/{userId}/{pokemonId}', [PokemonController::class, 'deleteFromFavorite']);
 
 Route::post('/users/add', [UserController::class, 'create']);
 Route::post('/users/sign-in', [UserController::class, 'checkIfUserExists']);
